Allow keypad key bindings to fail
The workstation binds actions to the arrow and home/end/etc keys, using
both the keypad (KP_) names and the non-keypad names.  On macos laptops,
these used to be accepted, but at least with macos 11.4, it now fails
because the KP_ names are not found.  If these keys do not exist, that
is fine, so we can guard the call with "catch" (noting that "try" is not
available until tcl 8.6).  It'd be nicer if we could test explicitly
whether the key exist, and skip the bind -- at the moment, if the bind
fails for an unexpected reason, we are silently ignoring it.